Proto heritage

+++ Projet2024-ejb/src/java/FACADE/ServiceFacade.java
@@ -8,6 +8,7 @@ import ENTITE.Service;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 
 /**
@@ -48,17 +49,29 @@ public class ServiceFacade extends AbstractFacade<Service> implements ServiceFac
     }
 
     @Override
-    public void supprimerService(Long id) {
+public void supprimerService(Long id) {
+    try {
+        if (id == null) {
+            throw new IllegalArgumentException("L'identifiant du service ne peut pas être null.");
+        }
         Service service = em.find(Service.class, id);
         if (service != null) {
             em.remove(service);
         }
+    } catch (IllegalArgumentException e) {
+        // Gérer l'exception lorsque id est null
+        System.err.println("Erreur : " + e.getMessage());
+    } catch (Exception e) {
+        // Gérer toute autre exception
+        System.err.println("Une erreur s'est produite lors de la suppression du service : " + e.getMessage());
     }
+}
+
 
     @Override
-    public Service trouverServiceParNom(String nom) {
+    public Service trouverServiceParId(Long id) {
         try {
-            return em.createQuery("SELECT s FROM Service s WHERE s.nom = :variable", Service.class).setParameter("variable", nom).getSingleResult();
+            return em.createQuery("SELECT s FROM Service s WHERE s.id = :variable", Service.class).setParameter("variable", id).getSingleResult();
         } catch (Exception e) {
             return null;
         }
@@ -68,7 +81,8 @@ public class ServiceFacade extends AbstractFacade<Service> implements ServiceFac
     public List<Service> trouverTousLesServices() {
         return em.createQuery("SELECT s FROM Service s", Service.class).getResultList();
     }
-    
+
+
